+++ packages/cli/src/commands/sandbox/sandbox-secret/sandbox_secret_set_command.test.ts
@@ -1,11 +1,13 @@
 import { beforeEach, describe, it, mock } from 'node:test';
-import { AmplifyPrompter } from '@aws-amplify/cli-core';
+import { AmplifyPrompter, printer } from '@aws-amplify/cli-core';
 import yargs, { CommandModule } from 'yargs';
 import { TestCommandRunner } from '../../../test-utils/command_runner.js';
 import assert from 'node:assert';
 import { SandboxBackendIdResolver } from '../sandbox_id_resolver.js';
 import { SecretIdentifier, getSecretClient } from '@aws-amplify/backend-secret';
 import { SandboxSecretSetCommand } from './sandbox_secret_set_command.js';
+import { ReadStream } from 'node:tty';
+import { PassThrough } from 'node:stream';
 
 const testSecretName = 'testSecretName';
 const testSecretValue = 'testSecretValue';
@@ -24,6 +26,7 @@ void describe('sandbox secret set command', () => {
     'setSecret',
     (): Promise<SecretIdentifier> => Promise.resolve(testSecretIdentifier)
   );
+  const printMock = mock.method(printer, 'print');
 
   const sandboxIdResolver: SandboxBackendIdResolver = {
     resolve: (identifier?: string) =>
@@ -34,9 +37,13 @@ void describe('sandbox secret set command', () => {
       }),
   } as SandboxBackendIdResolver;
 
+  const mockReadStream = new PassThrough();
+  (mockReadStream as unknown as ReadStream).isTTY = true; // Fake a TTY in tests
+
   const sandboxSecretSetCmd = new SandboxSecretSetCommand(
     sandboxIdResolver,
-    secretClient
+    secretClient,
+    mockReadStream as unknown as ReadStream
   );
 
   const parser = yargs().command(
@@ -47,6 +54,7 @@ void describe('sandbox secret set command', () => {
 
   beforeEach(async () => {
     secretSetMock.mock.resetCalls();
+    printMock.mock.resetCalls();
   });
 
   void it('sets a secret', async (contextual) => {
@@ -69,6 +77,10 @@ void describe('sandbox secret set command', () => {
       testSecretName,
       testSecretValue,
     ]);
+    assert.equal(
+      printMock.mock.calls[0].arguments[0],
+      `Successfully created version ${testSecretIdentifier.version} of secret ${testSecretIdentifier.name}`
+    );
   });
 
   void it('sets a secret in a named sandbox', async (contextual) => {
@@ -89,6 +101,41 @@ void describe('sandbox secret set command', () => {
         type: 'sandbox',
         namespace: testBackendId,
         name: 'anotherName',
+      },
+      testSecretName,
+      testSecretValue,
+    ]);
+  });
+
+  void it('sets a secret using redirection', async () => {
+    const readStream = new PassThrough(); // 'isTTY' flag doesn't exist on PassThrough stream.
+
+    const sandboxSecretSetCmdWithStream = new SandboxSecretSetCommand(
+      sandboxIdResolver,
+      secretClient,
+      readStream as unknown as ReadStream
+    );
+
+    const parserWithStream = yargs().command(
+      sandboxSecretSetCmdWithStream as unknown as CommandModule
+    );
+
+    const commandRunnerWithStream = new TestCommandRunner(parserWithStream);
+
+    const setCommandPromise = commandRunnerWithStream.runCommand(
+      `set ${testSecretName}`
+    );
+    readStream.write(testSecretValue);
+    readStream.end();
+    await setCommandPromise;
+
+    assert.equal(secretSetMock.mock.callCount(), 1);
+
+    assert.deepStrictEqual(secretSetMock.mock.calls[0].arguments, [
+      {
+        type: 'sandbox',
+        namespace: testBackendId,
+        name: testSandboxName,
       },
       testSecretName,
       testSecretValue,